+++ Forge/src/main/java/fuzs/puzzleslib/impl/event/ForgeEventInvokerRegistryImpl.java
@@ -431,6 +431,11 @@ public final class ForgeEventInvokerRegistryImpl implements ForgeEventInvokerReg
         INSTANCE.register(LivingEvents.Jump.class, LivingEvent.LivingJumpEvent.class, (LivingEvents.Jump callback, LivingEvent.LivingJumpEvent evt) -> {
             LivingJumpHelper.onLivingJump(callback, evt.getEntity());
         });
+        INSTANCE.register(LivingEvents.Visibility.class, LivingEvent.LivingVisibilityEvent.class, (LivingEvents.Visibility callback, LivingEvent.LivingVisibilityEvent evt) -> {
+            callback.onLivingVisibility(evt.getEntity(), evt.getLookingEntity(), MutableDouble.fromEvent(visibilityModifier -> {
+                evt.modifyVisibility(visibilityModifier / evt.getVisibilityModifier());
+            }, evt::getVisibilityModifier));
+        });
         if (ModLoaderEnvironment.INSTANCE.isClient()) {
             ForgeClientEventInvokers.register();
         }