+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog].
 ### Added
 - Added `MovementInputUpdateCallback`
 - Added `LivingEvents$Jump`
+- Added `LivingEvents$Visibility`
 
 ## [v7.0.9-1.20] - 2023-06-19
 ### Added